+++ packages/core/client/src/modules/blocks/data-blocks/grid-card/gridCardBlockSettings.ts
@@ -2,85 +2,23 @@ import { ArrayItems } from '@formily/antd-v5';
 import { ISchema } from '@formily/json-schema';
 import { useField, useFieldSchema } from '@formily/react';
 import _ from 'lodash';
-import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { SchemaSettings } from '../../../../application/schema-settings/SchemaSettings';
 import { useFormBlockContext } from '../../../../block-provider';
 import { useCollection_deprecated, useSortFields } from '../../../../collection-manager';
 import { removeNullCondition, useDesignable } from '../../../../schema-component';
-import {
-  defaultColumnCount,
-  gridSizes,
-  pageSizeOptions,
-  screenSizeMaps,
-  screenSizeTitleMaps,
-} from '../../../../schema-component/antd/grid-card/options';
+import { pageSizeOptions } from '../../../../schema-component/antd/grid-card/options';
 import { SchemaSettingsTemplate } from '../../../../schema-settings';
 import { SchemaSettingsDataScope } from '../../../../schema-settings/SchemaSettingsDataScope';
 import { setDataLoadingModeSettingsItem } from '../details-multi/setDataLoadingModeSettingsItem';
-import { columnCountMarks } from './utils';
+import { SetTheCountOfColumnsDisplayedInARow } from './SetTheCountOfColumnsDisplayedInARow';
 
 export const gridCardBlockSettings = new SchemaSettings({
   name: 'blockSettings:gridCard',
   items: [
     {
       name: 'SetTheCountOfColumnsDisplayedInARow',
-      type: 'modal',
-      useComponentProps() {
-        const { t } = useTranslation();
-        const fieldSchema = useFieldSchema();
-        const field = useField();
-        const { dn } = useDesignable();
-        const columnCount = field.decoratorProps.columnCount || defaultColumnCount;
-
-        const columnCountSchema = useMemo(() => {
-          return {
-            'x-component': 'Slider',
-            'x-decorator': 'FormItem',
-            'x-component-props': {
-              min: 1,
-              max: 24,
-              marks: columnCountMarks,
-              tooltip: {
-                formatter: (value) => `${value}${t('Column')}`,
-              },
-              step: null,
-            },
-          };
-        }, [t]);
-
-        const columnCountProperties = useMemo(() => {
-          return gridSizes.reduce((o, k) => {
-            o[k] = {
-              ...columnCountSchema,
-              title: t(screenSizeTitleMaps[k]),
-              description: `${t('Screen size')} ${screenSizeMaps[k]} ${t('pixels')}`,
-            };
-            return o;
-          }, {});
-        }, [columnCountSchema, t]);
-
-        return {
-          title: t('Set the count of columns displayed in a row'),
-          initialValues: columnCount,
-          schema: {
-            type: 'object',
-            title: t('Set the count of columns displayed in a row'),
-            properties: columnCountProperties,
-          } as ISchema,
-
-          onSubmit: (columnCount) => {
-            _.set(fieldSchema, 'x-decorator-props.columnCount', columnCount);
-            field.decoratorProps.columnCount = columnCount;
-            dn.emit('patch', {
-              schema: {
-                ['x-uid']: fieldSchema['x-uid'],
-                'x-decorator-props': fieldSchema['x-decorator-props'],
-              },
-            });
-          },
-        };
-      },
+      Component: SetTheCountOfColumnsDisplayedInARow,
     },
     {
       name: 'SetTheDataScope',
